// frontend/src/types/masterdata.ts
export interface CategoryMasterDataTypes {
    id: string;
    name: string;
    createdAt: string;
    updatedAt: string;
}

export interface IndustryMasterDataTypes {
    id: string;
    name: string;
    createdAt: string;
    updatedAt: string;
}

export interface UnitMasterDataTypes {
    id: string;
    name: string;
    symbol?: string;
    createdAt: string;
    updatedAt: string;
}